refactor(mock-api): split into domain modules and fill e2e route gaps
The 1500-line scripts/mock-api-core.mjs is now a thin re-export shim over
scripts/mock-api/, organized by domain:

  state.mjs       — mockBehavior, request log, tunnels, user/team factories
  http.mjs        — CORS, json/html senders, body parsing, header redaction
  socket.mjs      — engine.io polling + WebSocket upgrade
  admin.mjs       — /__admin/{health,requests,behavior,reset}
  server.mjs      — http server lifecycle + route dispatch
  routes/
    auth.mjs           — /auth/*, /telegram/login-tokens/*
    user.mjs           — /settings, /teams*, /rewards/me, /referral/*, /users
    invites.mjs        — /invite/*
    payments.mjs       — /payments/*, /billing/*
    integrations.mjs   — /agent-integrations/*, /gmail/*, /notion/*, /telegram/*, /openai/v1/*
    webhooks.mjs       — /webhooks/core/*, /webhooks/ingress/*
    conversations.mjs  — /conversations, /messages*, /channels*, /notifications
    version.mjs        — /version-check
    cron.mjs           — /settings/cron-jobs, /settings/webhooks-triggers

Gap-fill stubs added for routes the app/core call but the mock never
implemented: /version-check, /auth/email/send-link, /auth/telegram,
/users, /teams/join, /conversations*, /messages*, /channels*,
/notifications, /webhooks/ingress*, /settings/cron-jobs*,
/settings/webhooks-triggers*, /payments/credits/{top-up,transactions,
auto-recharge/cards/setup-intent,auto-recharge/cards/:id},
/agent-integrations/composio/{github/repos,tools}, /agent-integrations/apify/runs/*.

All 2042 vitest unit tests pass; manual curl smoke against the split
server confirms admin endpoints, existing routes, and new gap routes
all respond. Existing import paths (scripts/mock-api-server.mjs,
app/test/e2e/mock-server.ts) keep working unchanged.
